MC-214: Add a filter to get exportable products from variant group
