+++ Resources/config/processors.yml
@@ -42,6 +42,7 @@ services:
             - '@pim_magento_connector.manager.attribute'
             - '@pim_magento_connector.manager.association_type'
             - '@pim_magento_connector.manager.group'
+            - '@pim_magento_connector.filters.exportable_product'
 
     pim_magento_connector.processor.product_association:
         class: %pim_magento_connector.processor.product_association.class%
